Add groups and use them for ownership of objects and storage units

+++ PiratenKarte/Server/Controllers/MapObjectsController.cs
@@ -8,7 +8,7 @@ using PiratenKarte.Shared.RequestModels;
 
 namespace PiratenKarte.Server.Controllers;
 
-public class MapObjectsController : CrudController<MapObject, DAL.Models.MapObject> {
+public class MapObjectsController : CrudController<MapObjectDTO, DAL.Models.MapObject> {
 	protected override RepositoryBase<DAL.Models.MapObject> Repository => DB.MapObjectRepo;
 
 	public override string PermissionBaseName => "objects";
@@ -16,42 +16,96 @@ public class MapObjectsController : CrudController<MapObject, DAL.Models.MapObje
     public MapObjectsController(DB db, IMapper mapper) : base(db, mapper) { }
 
     [HttpGet]
-	[Permission("objects_read")]
-    public IEnumerable<MapObject> GetMap() => DB.MapObjectRepo.GetMap().Select(Mapper.Map<MapObject>);
+    [Permission("objects_read")]
+    public IActionResult GetMap() {
+		if (!TryGetUser(out var user))
+			return BadRequest();
 
-    [HttpPost]
-    public override Guid Create(MapObject item) {
-		throw new NotSupportedException();
+		var onMap = DB.MapObjectRepo.GetMap();
+		var filtered = new List<MapObjectDTO>();
+
+		foreach (var mo in onMap) {
+			if (user.GroupIds.Contains(mo.GroupId))
+				filtered.Add(Mapper.Map<MapObjectDTO>(mo));
+		}
+
+		return Ok(filtered);
     }
+
+    [NonAction]
+    public override PagedData<MapObjectDTO> GetPaged(int page, int itemsPerPage) => throw new NotImplementedException();
+
+    public IActionResult GetPagedEx(int page, int itemsPerPage) {
+        if (!TryGetUser(out var user))
+            return BadRequest();
+
+        var pagedData = new PagedData<MapObjectDTO>() {
+            Data = [],
+            TotalCount = 0
+        };
+
+        pagedData.Data = DB.MapObjectRepo.GetAll()
+            .Where(mo => user.GroupIds.Contains(mo.GroupId))
+            .Skip(page * itemsPerPage).Take(itemsPerPage)
+            .Select(Mapper.Map<MapObjectDTO>).ToList();
+        pagedData.TotalCount = DB.MapObjectRepo.CountVisible(user);
+
+        return Ok(pagedData);
+    }
+
+    [NonAction]
+    public override Guid Create(MapObjectDTO item) => throw new NotSupportedException();
 
     [HttpPost]
     [Permission("objects_create")]
-    public Guid CreateSingle(CreateNewObject request) {
+    public IActionResult CreateSingle(CreateNewObject request) {
+		if (!TryGetUser(out var user))
+			return BadRequest();
+
         var obj = Mapper.Map<DAL.Models.MapObject>(request.Object);
 		obj.Storage = request.StorageId == null ? null : DB.StorageDefinitionRepo.Get(request.StorageId.Value);
 
-		return DB.MapObjectRepo.Insert(obj);
+		if (!IsUserInGroup(user, obj.GroupId))
+			return Unauthorized();
+
+		return Ok(DB.MapObjectRepo.Insert(obj));
     }
 
 	[HttpPost]
     [Permission("objects_create")]
-    public void CreateMany(CreateNewObjectBulk request) {
-		var storage = request.StorageId == null ? null : DB.StorageDefinitionRepo.Get(request.StorageId.Value);
+    public IActionResult CreateMany(CreateNewObjectBulk request) {
+        if (!TryGetUser(out var user))
+            return BadRequest();
+        if (!IsUserInGroup(user, request.GroupId))
+            return Unauthorized();
 
+        var storage = request.StorageId == null ? null : DB.StorageDefinitionRepo.Get(request.StorageId.Value);
 		for (int i = 0; i < request.Count; i++) {
-			var obj = Mapper.Map<DAL.Models.MapObject>(request.Template);
-			obj.Name = string.Format(obj.Name, i);
+            var obj = Mapper.Map<DAL.Models.MapObject>(request.Template);
+            obj.Name = string.Format(obj.Name, i);
 			obj.Storage = storage;
 			DB.MapObjectRepo.Insert(obj);
 		}
+
+		return Ok();
 	}
 
 	[HttpPost]
     [Permission("objects_delete")]
-    public void DeleteMany([FromBody] IEnumerable<Guid> ids) {
-		foreach (var id in ids)
-			DB.MapObjectRepo.Delete(id);
-	}
+    public IActionResult DeleteMany([FromBody] IEnumerable<Guid> ids) {
+        if (!TryGetUser(out var user))
+            return BadRequest();
+
+        foreach (var id in ids) {
+            var obj = DB.MapObjectRepo.Get(id);
+            if (!IsUserInGroup(user, obj.GroupId))
+                continue;
+
+            DB.MapObjectRepo.Delete(id);
+        }
+
+        return Ok();
+    }
 
 	[HttpPost]
     [Permission("objects_comments_create")]
@@ -68,8 +122,7 @@ public class MapObjectsController : CrudController<MapObject, DAL.Models.MapObje
     [Permission("objects_comments_delete")]
     public void DeleteComment(DeleteObjectComment request) {
 		var obj = DB.MapObjectRepo.Get(request.ObjectId);
-        if (obj.Comments == null)
-            obj.Comments = new List<DAL.Models.ObjectComment>();
+        obj.Comments ??= [];
 
 		obj.Comments.RemoveAll(c => c.Id == request.CommentId);
 		DB.MapObjectRepo.Update(obj);
@@ -77,34 +130,56 @@ public class MapObjectsController : CrudController<MapObject, DAL.Models.MapObje
 
 	[HttpPost]
     [Permission("objects_update")]
-    public void SetStorage(SetObjectStorage request) {
-		var obj = DB.MapObjectRepo.Get(request.ObjectId);
+    public IActionResult SetStorage(SetObjectStorage request) {
+        if (!TryGetUser(out var user))
+            return BadRequest();
+
+        var obj = DB.MapObjectRepo.Get(request.ObjectId);
+        if (!IsUserInGroup(user, obj.GroupId))
+            return Unauthorized();
 
         obj.Storage = request.StorageId == null ? null : DB.StorageDefinitionRepo.Get(request.StorageId.Value);
 
         DB.MapObjectRepo.Update(obj);
+		return Ok();
     }
 
 	[HttpPost]
     [Permission("objects_update")]
-    public void SetStorageMany(SetObjectStorageMany request) {
-		var storage = request.StorageId == null ? null : DB.StorageDefinitionRepo.Get(request.StorageId.Value);
+    public IActionResult SetStorageMany(SetObjectStorageMany request) {
+        if (!TryGetUser(out var user))
+            return BadRequest();
 
-		foreach (var objId in request.ObjectIds) {
+        var storage = request.StorageId == null ? null : DB.StorageDefinitionRepo.Get(request.StorageId.Value);
+        if (storage == null || !IsUserInGroup(user, storage.GroupId))
+            return Unauthorized();
+
+        foreach (var objId in request.ObjectIds) {
 			var obj = DB.MapObjectRepo.Get(objId);
+			if (!IsUserInGroup(user, obj.GroupId))
+				continue;
+
 			obj.Storage = storage;
 			DB.MapObjectRepo.Update(obj);
 		}
+
+		return Ok();
 	}
 
 	[HttpPost]
     [Permission("objects_update")]
-    public void UpdatePosition(SetObjectPosition request) {
-		var obj = DB.MapObjectRepo.Get(request.ObjectId);
+    public IActionResult UpdatePosition(SetObjectPosition request) {
+        if (!TryGetUser(out var user))
+            return BadRequest();
+
+        var obj = DB.MapObjectRepo.Get(request.ObjectId);
+		if (!IsUserInGroup(user, obj.GroupId))
+			return Unauthorized();
 
 		obj.Storage = null;
 		obj.LatLon = Mapper.Map<DAL.Models.LatitudeLongitude>(request.Position);
 
 		DB.MapObjectRepo.Update(obj);
+		return Ok();
 	}
 }
